requirements.txt
streamlit
yt-dlp
moviepy
whisper-openai

